requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bitarray==2.5.1
black==19.10b0
boto3==1.17.33
botocore==1.20.33
brotlipy==0.7.0
cachetools==5.2.0
certifi==2020.12.5
cfgv==3.2.0
click==8.0.0a1
cloudpickle==1.6.0
colorama==0.4.4
configparser==5.0.2
coverage==5.5
cycler==0.11.0
Cython==0.29.22
dask==2.30.0
dask-jobqueue==0.7.0
dataclasses==0.6
decorator==4.4.2
distlib==0.3.1
distributed==2.30.1
docker-pycreds==0.4.0
editdistance==0.5.3
entrypoints==0.3
fairscale==0.4.6
fairtask==1.1.0
fairtask-slurm==0.3.1
fastBPE==0.1.0
filelock==3.0.12
flake8==3.7.9
fonttools==4.34.4
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.3
HeapDict==1.0.1
hydra-core==1.0.7
hydra-fair-cluster==1.0.0
hydra-fair-plugins==1.0.0
hydra-fairtask==1.0.0
hydra-submitit-launcher==1.1.0
identify==2.2.0
importlib-resources==5.1.2
iopath==0.1.6
ipdb==0.13.7
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.18.0
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.4.4
Markdown==3.3.7
matplotlib==3.5.2
mccabe==0.6.1
mkl-fft==1.3.0
more-itertools==8.7.0
mpi4py==3.0.3
msgpack==1.0.2
multidict==5.1.0
mypy==0.790
mypy-extensions==0.4.3
nodeenv==1.5.0
numpy==1.23.5
oauthlib==3.2.0
omegaconf==2.0.6
packaging==20.9
pandas==1.4.3
parameterized==0.8.1
parso==0.8.1
pathspec==0.8.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pluggy==0.13.1
portalocker==2.0.0
pre-commit==2.11.1
promise==2.3
prompt-toolkit==3.0.17
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pydantic==1.8.1
pyflakes==2.1.1
Pygments==2.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.10.0
pytest-mpi==0.4
pytest-timeout==1.4.2
python-dateutil==2.8.1
pytz==2022.1
PyYAML==5.4.1
regex==2021.3.17
remote-pdb==2.1.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.3.6
sacrebleu @ git+https://github.com/mjpost/sacrebleu.git@8e7abf591f6acff69dcbcad8d7af94754296fcb4
sacremoses==0.0.43
scipy==1.8.1
seaborn==0.11.2
sentencepiece==0.1.95
sentry-sdk==1.0.0
shortuuid==1.0.1
smmap==3.0.5
sortedcontainers==2.3.0
submitit==1.3.0
subprocess32==3.5.4
subword-nmt==0.3.7
tabulate==0.8.9
tblib==1.7.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
toml==0.10.2
toolz==0.11.1
torch==1.8.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typed-ast==1.4.2
virtualenv==20.4.3
wandb==0.10.22
wcwidth==0.2.5
Werkzeug==2.1.2
yarl==1.6.3
zict==2.0.0
zipp==3.4.1
